requirements.txt
aiogram==3.13.1
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1
